Store existing atags address at startup if chainloading.
This patch stores the atag address at startup. It also stores the address
which u-boot was originally loaded to. These two pieces of information are
REQUIRED to successfully boot the uImage appended to u-boot.bin at offset
0x40000.

The atag address is required in order to pass through the atags set by NOLO.
NOLO is a proprietary bootloader, therefore we don't know what atags it can
set. The only option is to back up the atags which it sets at runtime. Since
this information is passed through r2, it is lost as soon as u-boot begins
execution. Therefore it MUST be backed up before any other code executes.

The load address of u-boot.bin is required because without it we do not know
where the uImage was loaded. This information can only be obtained before
u-boot is copied to _TEXT_BASE since that moves the location of _start.

Signed-off-by: Alistair Buxton <a.j.buxton@gmail.com>

Conflicts:

	arch/arm/cpu/armv7/start.S