Make bootm optionally use pre-existing atags for Linux kernel boot.
This patch adapts the bootm command so that it can use an existing atags command
set up by a previous bootloader. If the environment variable "atags" is unset,
bootm behaves as normal. If "atags" is set, bootm will skip all boot args setup
entirely, and pass the address found in "atags". For example, if a previous boot
loader already set up the atags struct at 0x80000100:

setenv atags 0x80000100; bootm 0x80008000

This is intended to be used in combination with the previous patch to pass
atags set by the previous bootloader.

Signed-off-by: Alistair Buxton <a.j.buxton@gmail.com>